requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.30.0
python-dotenv>=1.0.0
databento>=0.40.0
py_vollib_vectorized>=0.1.1
scipy>=1.10.0
